Move node replacement system into its own module
